+++ client/src/EglData.hpp
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: MIT
 /*!
  * @file
- * @brief Very simple GLES3 renderer for WebRTC client.
- * @author Moshi Turner <moses@collabora.com>
+ * @brief Encapsulate EGL setup/teardown.
  * @author Ryan Pavlik <rpavlik@collabora.com>
  */
 
@@ -50,44 +49,4 @@ struct EglData
 	EGLContext context = EGL_NO_CONTEXT;
 	EGLSurface surface = EGL_NO_SURFACE;
 	EGLConfig config = nullptr;
-};
-
-class Renderer
-{
-public:
-	Renderer() = default;
-	~Renderer();
-
-	Renderer(const Renderer &) = delete;
-	Renderer(Renderer &&) = delete;
-
-	Renderer &
-	operator=(const Renderer &) = delete;
-	Renderer &
-	operator=(Renderer &&) = delete;
-
-	/// Create resources. Must call with EGL Context current
-	void
-	setupRender();
-
-	/// Destroy resources. Must call with EGL context current.
-	void
-	reset();
-
-	/// Draw texture to framebuffer. Must call with EGL Context current.
-	void
-	draw(GLuint texture, GLenum texture_target) const;
-
-
-private:
-	void
-	setupShaders();
-	void
-	setupQuadVertexData();
-
-	GLuint program = 0;
-	GLuint quadVAO = 0;
-	GLuint quadVBO = 0;
-
-	GLint textureSamplerLocation_ = 0;
 };
